+++ screens/TabOneScreen.tsx
@@ -1,5 +1,5 @@
-import React, { useState } from 'react';
-import { StyleSheet } from 'react-native';
+import React, { useRef, useState } from 'react';
+import { Button, StyleSheet } from 'react-native';
 import ProgressBar from '../components/ProgressBar';
 import EditScreenInfo from '../components/EditScreenInfo';
 import { Text, View } from '../components/Themed';
@@ -22,9 +22,27 @@ const styles = StyleSheet.create({
 });
 
 const TabOneScreen = () => {
+    const [currentStep, setNextStep] = useState(0);
+    const maxSteps = 5;
+    const stepSize = 1 / maxSteps;
+
+    function handleStepNext() {
+        if (currentStep + stepSize <= 1) {
+            setNextStep(currentStep + stepSize);
+        }
+    }
+
+    function handleStepPrevious() {
+        if (currentStep - stepSize >= 0) {
+            setNextStep(currentStep - stepSize);
+        }
+    }
+
     return (
         <View style={styles.container}>
-            <ProgressBar maxSteps={5} />
+            <ProgressBar maxSteps={maxSteps} currentStep={currentStep} />
+            <Button onPress={handleStepNext} title="Next" />
+            <Button onPress={handleStepPrevious} title="Previous" />
         </View>
     );
 };
